Streamlit/requirements.txt
streamlit
requests
opencensus-ext-azure
opencensus-ext-requests
azure-monitor-opentelemetry
dotenv